@@ -12,11 +12,11 @@
   "dependencies": {
     "@babel/runtime": "^7.2.0",
     "body-parser": "^1.18.3",
-    "dayjs": "^1.8.5",
+    "dayjs": "^1.8.6",
     "express": "^4.16.4",
     "material-design-icons": "^3.0.1",
-    "react": "^16.8.1",
-    "react-dom": "^16.8.1",
+    "react": "^16.8.2",
+    "react-dom": "^16.8.2",
     "react-hot-loader": "^4.6.5",
     "react-redux": "^6.0.0",
     "react-router": "^4.3.1",
@@ -32,25 +32,25 @@
     "xmlhttprequest": "^1.8.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/core": "^7.3.3",
+    "@babel/plugin-proposal-class-properties": "^7.3.3",
     "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-transform-runtime": "^7.2.0",
     "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
+    "@babel/preset-typescript": "^7.3.3",
     "@storybook/addon-actions": "^4.1.7",
     "@storybook/react": "^4.1.7",
     "@types/express": "^4.16.0",
-    "@types/jest": "^24.0.0",
-    "@types/react": "^16.8.2",
-    "@types/react-dom": "^16.8.0",
+    "@types/jest": "^24.0.5",
+    "@types/react": "^16.8.3",
+    "@types/react-dom": "^16.8.1",
     "@types/react-redux": "^7.0.0",
-    "@types/react-router": "^4.4.3",
+    "@types/react-router": "^4.4.4",
     "@types/react-router-dom": "^4.3.1",
-    "@types/storybook__addon-actions": "^3.4.1",
-    "@types/storybook__react": "^4.0.0",
+    "@types/storybook__addon-actions": "^3.4.2",
+    "@types/storybook__react": "^4.0.1",
     "@typescript-eslint/eslint-plugin": "^1.3.0",
     "@typescript-eslint/parser": "^1.3.0",
     "babel-core": "^7.0.0-bridge.0",
@@ -61,12 +61,13 @@
     "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.0",
-    "eslint": "^5.13.0",
+    "eslint": "^5.14.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-loader": "^2.1.2",
     "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jest": "^22.2.2",
+    "eslint-plugin-jest": "^22.3.0",
     "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react-hooks": "^1.0.2",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.1.0",
     "mini-css-extract-plugin": "^0.5.0",
@@ -74,7 +75,7 @@
     "null-loader": "^0.1.1",
     "postcss-loader": "^3.0.0",
     "prettier-stylelint": "^0.4.2",
-    "react-test-renderer": "^16.8.1",
+    "react-test-renderer": "^16.8.2",
     "redux-logger": "^3.0.6",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
@@ -84,8 +85,8 @@
     "stylelint-selector-bem-pattern": "^2.0.0",
     "stylelint-webpack-plugin": "^0.10.5",
     "typescript": "^3.3.3",
-    "webpack": "^4.29.3",
-    "webpack-bundle-analyzer": "^3.0.3",
+    "webpack": "^4.29.4",
+    "webpack-bundle-analyzer": "^3.0.4",
     "webpack-cli": "^3.2.3",
     "webpack-dev-server": "^3.1.14",
     "webpack-manifest-plugin": "^2.0.4",
