@@ -6,7 +6,7 @@ import {
 } from 'redux-observable';
 import { BehaviorSubject } from 'rxjs';
 import { AjaxCreationMethod } from 'rxjs/internal/observable/dom/AjaxObservable';
-import { mergeMap } from 'rxjs/operators';
+import { concatMap } from 'rxjs/operators';
 import { epics } from './dummy';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
@@ -19,7 +19,7 @@ export default () => {
       action$: ActionsObservable<any>,
       state$: StateObservable<any>,
       { ajax }: { ajax: AjaxCreationMethod }
-    ) => epic$.pipe<any>(mergeMap(epic => epic(action$, state$, { ajax }))),
+    ) => epic$.pipe<any>(concatMap(epic => epic(action$, state$, { ajax }))),
     epicSubject$: epic$
   };
 };
