@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useCallback, useState, useEffect } from 'react';
 import dayjs from 'dayjs';
 // @ts-ignore
 import relativeTime from 'dayjs/plugin/relativeTime';
@@ -14,67 +14,60 @@ export type Props = {
   onLike: (book: Book) => void;
 };
 
-export default class BookCard extends Component<Props> {
-  constructor(props: Props) {
-    super(props);
-    dayjs.extend(relativeTime);
-  }
+dayjs.extend(relativeTime);
 
-  handleViewDetails = () => {
-    this.props.onViewDetails(this.props.book);
-  };
+const BookCard = (props: Props) => {
+  const handleViewDetails = useCallback(() => {
+    props.onViewDetails(props.book);
+  }, [props.onViewDetails, props.book]);
 
-  handleLike = () => {
-    this.props.onLike(this.props.book);
-  };
+  const handleLike = useCallback(() => {
+    props.onLike(props.book);
+  }, [props.onLike, props.book]);
 
-  // @ts-ignore
-  getRelativeDate = () => dayjs(this.props.book.published).fromNow();
+  const [relativeDate, setRelativeDate] = useState('');
+  useEffect(() => {
+    // @ts-ignore
+    setRelativeDate(dayjs(props.book.published).fromNow());
+  }, [props.book.published]);
 
-  shouldComponentUpdate(nextProps: Props) {
-    return nextProps.book !== this.props.book;
-  }
-
-  render() {
-    return (
-      <div className="book-card">
-        <div className="book-card__image" onClick={this.handleViewDetails}>
-          <img src={this.props.book.cover} alt={this.props.book.name} />
-        </div>
-        <div className="book-card__header" onClick={this.handleViewDetails}>
-          <div className="book-card__title">{this.props.book.name}</div>
-          <div className="book-card__author">{this.props.book.author.name}</div>
-        </div>
-        <div className="book-card__actions">
-          <div className="book-card__action" onClick={this.handleLike}>
-            {this.props.book.liked && (
-              <svg viewBox={favoriteIcon.viewBox} className="book-card__icon">
-                <use xlinkHref={`#${favoriteIcon.id}`} />
-              </svg>
-            )}
-            {!this.props.book.liked && (
-              <svg
-                viewBox={favoriteBorderIcon.viewBox}
-                className="book-card__icon"
-              >
-                <use xlinkHref={`#${favoriteBorderIcon.id}`} />
-              </svg>
-            )}
-            {`${this.props.book.likes} ${
-              this.props.book.likes > 1 ? 'likes' : 'like'
-            }`}
-          </div>
-          <div className="book-card__action book-card__action--align-right book-card__action--no-click">
+  return (
+    <div className="book-card">
+      <div className="book-card__image" onClick={handleViewDetails}>
+        <img src={props.book.cover} alt={props.book.name} />
+      </div>
+      <div className="book-card__header" onClick={handleViewDetails}>
+        <div className="book-card__title">{props.book.name}</div>
+        <div className="book-card__author">{props.book.author.name}</div>
+      </div>
+      <div className="book-card__actions">
+        <div className="book-card__action" onClick={handleLike}>
+          {props.book.liked && (
+            <svg viewBox={favoriteIcon.viewBox} className="book-card__icon">
+              <use xlinkHref={`#${favoriteIcon.id}`} />
+            </svg>
+          )}
+          {!props.book.liked && (
             <svg
-              viewBox={calendarIcon.viewBox}
+              viewBox={favoriteBorderIcon.viewBox}
               className="book-card__icon"
             >
-              <use xlinkHref={`#${calendarIcon.id}`} />
+              <use xlinkHref={`#${favoriteBorderIcon.id}`} />
             </svg>
-            <div>{this.getRelativeDate()}</div>
-          </div>
+          )}
+          {`${props.book.likes} ${props.book.likes > 1 ? 'likes' : 'like'}`}
+        </div>
+        <div className="book-card__action book-card__action--align-right book-card__action--no-click">
+          <svg viewBox={calendarIcon.viewBox} className="book-card__icon">
+            <use xlinkHref={`#${calendarIcon.id}`} />
+          </svg>
+          <div>{relativeDate}</div>
         </div>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
+
+BookCard.displayName = 'BookCard';
+
+export default React.memo(BookCard);
